move cost functions to cost function file
